move node to electron-main, leaving react stuff in contrib/
