+++ package.json
@@ -75,6 +75,8 @@
     "@microsoft/1ds-core-js": "^3.2.13",
     "@microsoft/1ds-post-js": "^3.2.13",
     "@parcel/watcher": "2.1.0",
+    "@rrweb/record": "^2.0.0-alpha.17",
+    "@rrweb/types": "^2.0.0-alpha.17",
     "@vscode/deviceid": "^0.1.1",
     "@vscode/iconv-lite-umd": "0.7.0",
     "@vscode/policy-watcher": "^1.1.4",
@@ -109,6 +111,7 @@
     "native-watchdog": "^1.4.1",
     "node-pty": "1.1.0-beta21",
     "open": "^8.4.2",
+    "posthog-node": "^4.3.1",
     "tas-client-umd": "0.2.0",
     "v8-inspect-profiler": "^0.1.1",
     "vscode-oniguruma": "1.7.0",
@@ -218,7 +221,6 @@
     "path-browserify": "^1.0.1",
     "postcss": "^8.4.33",
     "postcss-nesting": "^12.0.2",
-    "posthog-js": "^1.184.2",
     "pump": "^1.0.1",
     "rcedit": "^1.1.0",
     "react": "^18.3.1",
