+++ src/vs/platform/void/electron-main/llmMessage/sendLLMMessage.ts
@@ -1,5 +1,6 @@
-import { posthog } from 'posthog-js'
-import type { OnText, OnError, OnFinalMessage, SendLLMMMessageParams, } from '../../../../../../../platform/void/common/llmMessageTypes.js';
+import { SendLLMMMessageParams, OnText, OnFinalMessage, OnError } from '../../common/llmMessageTypes.js';
+import { IMetricsService } from '../../common/metricsService.js';
+
 import { sendAnthropicMsg } from './anthropic.js';
 import { sendGeminiMsg } from './gemini.js';
 import { sendGreptileMsg } from './greptile.js';
@@ -8,7 +9,14 @@ import { sendOllamaMsg } from './ollama.js';
 import { sendOpenAIMsg } from './openai.js';
 
 
-export const sendLLMMessage = ({ messages, onText: onText_, onFinalMessage: onFinalMessage_, onError: onError_, abortRef: abortRef_, voidConfig, logging: { loggingName } }: SendLLMMMessageParams) => {
+export const sendLLMMessage = (
+	{
+		messages, onText: onText_, onFinalMessage: onFinalMessage_, onError: onError_,
+		abortRef: abortRef_, voidConfig, logging: { loggingName },
+	}: SendLLMMMessageParams,
+
+	metricsService: IMetricsService
+) => {
 	if (!voidConfig) return;
 
 	// trim message content (Anthropic and other providers give an error if there is trailing whitespace)
@@ -16,7 +24,7 @@ export const sendLLMMessage = ({ messages, onText: onText_, onFinalMessage: onFi
 
 	// only captures number of messages and message "shape", no actual code, instructions, prompts, etc
 	const captureChatEvent = (eventId: string, extras?: object) => {
-		posthog.capture(eventId, {
+		metricsService.capture(eventId, {
 			whichApi: voidConfig.default['whichApi'],
 			numMessages: messages?.length,
 			messagesShape: messages?.map(msg => ({ role: msg.role, length: msg.content.length })),
