+++ src/vs/platform/void/electron-main/llmMessageChannel.ts
@@ -3,17 +3,16 @@
  *  Void Editor additions licensed under the AGPLv3 License.
  *--------------------------------------------------------------------------------------------*/
 
-// this channel is registered in `app.ts`
-// code convention is to make a service responsible for this stuff, and not a channel, but this is simpler.
-// you could create one instance in electron-main/my-service.ts and one in browser/my-service.ts (and define the interface IMyService in common/my-service.ts), but we just use a channel here
-// registerSingleton(ISendLLMMessageService, SendLLMMessageService, InstantiationType.Delayed);
+// registered in app.ts
+// code convention is to make a service responsible for this stuff, and not a channel, but having fewer files is simpler...
 
 import { IServerChannel } from '../../../base/parts/ipc/common/ipc.js';
 import { Emitter, Event } from '../../../base/common/event.js';
-import { sendLLMMessage } from '../../../workbench/contrib/void/browser/react/out/sendLLMMessage/sendLLMMessage.js';
 import { listenerNames, ProxyOnTextPayload, ProxyOnErrorPayload, ProxyOnFinalMessagePayload, ProxyLLMMessageParams, AbortRef, SendLLMMMessageParams, ProxyLLMMessageAbortParams } from '../common/llmMessageTypes.js';
+import { sendLLMMessage } from './llmMessage/sendLLMMessage.js'
+import { IMetricsService } from '../common/metricsService.js';
 
-// NODE IMPLEMENTATION OF SENDLLMMESSAGE - calls sendLLMMessage() and returns listeners
+// NODE IMPLEMENTATION - calls actual sendLLMMessage() and returns listeners to it
 
 export class LLMMessageChannel implements IServerChannel {
 	private readonly _onText = new Emitter<ProxyOnTextPayload>();
@@ -29,7 +28,9 @@ export class LLMMessageChannel implements IServerChannel {
 	private readonly _abortRefOfRequestId: Record<string, AbortRef> = {}
 
 
-	constructor() { }
+	constructor(
+		private readonly metricsService: IMetricsService
+	) { }
 
 	// browser uses this to listen for changes
 	listen(_: unknown, event: typeof listenerNames[number]): Event<any> {
@@ -80,7 +81,7 @@ export class LLMMessageChannel implements IServerChannel {
 			onError: ({ error }) => { this._onError.fire({ requestId, error }); },
 			abortRef: this._abortRefOfRequestId[requestId],
 		}
-		sendLLMMessage(mainThreadParams);
+		sendLLMMessage(mainThreadParams, this.metricsService);
 	}
 
 	private _callAbort(params: ProxyLLMMessageAbortParams) {
