--- src/vs/workbench/contrib/void/browser/react/src/util/posthog.tsx
@@ -1,4 +1,0 @@
-
-import posthog from 'posthog-js';
-
-export { posthog }
